requirements.txt
# python 3.12

# application server
fastapi[standard]==0.115.5
uvicorn==0.32.1
gunicorn==21.2.*
starlette==0.41.3
httpx==0.27.2
python-jose==3.3.0

#database handling
sqlalchemy==2.0.36
sqlalchemy_utils==0.41.2
sqlmodel==0.0.22

# Data models/handling
pydantic==2.7.*
pandas==2.2.3
ulid-py==1.1.0

# configuration handling
fire==0.7.0
dynaconf==3.2.6

# calculation
scipy==1.14.1
numpy==2.1.3
biopython==1.84

# other
cachetools==5.5.0
slowapi==0.1.9